chore: move the tests
