+++ projects/Mallard/src/push-notifications/__tests__/push-notifications.spec.ts
@@ -1,4 +1,4 @@
-import { maybeRegister } from '../../push-notifications/push-notifications'
+import { maybeRegister } from '../push-notifications'
 import {
     getMockAsyncCache,
     getMockPromise,
